Updates to type files - Export MinimalTokenInfo and MinimalTokenMap   * Allows us to send less data to frontend - Move more types into specific type files - Reorganize imports
