+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from "./coingecko";
 export * from "./marketdata";
 export * from "./serum";
 export * from "./star-atlas";
+export * from "./tokens";
